azure_fn/requirements.txt
azure-functions==1.21.3

